% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_example_file.R
\name{get_example_file}
\alias{get_example_file}
\title{Get the file path to the example transcript within the package}
\usage{
get_example_file()
}
\value{
A file path
}
\description{
Get the file path to the example transcript within the package
}
\examples{
\dontrun{
get_example_file(dat)
}
}
