% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_act.R
\name{scrape_act}
\alias{scrape_act}
\title{Scrape ACT scores from a student's transcript}
\usage{
scrape_act(.data)
}
\arguments{
\item{.data}{A character vector returned from the read_transcript function.}
}
\value{
A tibble of ACT test scores
}
\description{
Scrape ACT scores from a student's transcript
}
\examples{
\dontrun{
scrape_act(dat)
}
}
