% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_all_tests.R
\name{scrape_all_tests}
\alias{scrape_all_tests}
\title{Scrape all available test scores from a student's transcript}
\usage{
scrape_all_tests(.data)
}
\arguments{
\item{.data}{A character vector returned from the read_transcript function.}
}
\value{
A tibble of available test scores
}
\description{
Scrape all available test scores from a student's transcript
}
\examples{
\dontrun{
scrape_all_tests(dat)
}
}
