% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_earned_credits.R
\name{scrape_earned_credits}
\alias{scrape_earned_credits}
\title{Scrape a student's credits earned while at BYU-I from their transcript}
\usage{
scrape_earned_credits(.data, tibble = TRUE)
}
\arguments{
\item{.data}{A character vector returned from the read_transcript function.}

\item{tibble}{A logical, specifying if a tibble should be returned or not.}
}
\value{
A tibble by default, otherwise a double vector.
}
\description{
Scrape a student's credits earned while at BYU-I from their transcript
}
\examples{
scrape_earned_credits(dat)
}
