% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_grades.R
\name{scrape_grades}
\alias{scrape_grades}
\title{Scrape a student's grades earned at BYUI from their transcript}
\usage{
scrape_grades(.data)
}
\arguments{
\item{.data}{A character vector returned from the read_transcript function.}
}
\value{
A tibble with four resulting columns: course, credit, grade & semester
}
\description{
Scrape a student's grades earned at BYUI from their transcript
}
\examples{
\dontrun{
scrape_grades(dat)
}
}
