% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_major.R
\name{scrape_major}
\alias{scrape_major}
\title{Scrape a student's major from their transcript}
\usage{
scrape_major(.data, tibble = TRUE)
}
\arguments{
\item{.data}{A character vector returned from the read_transcript function.}

\item{tibble}{A logical, specifying if a tibble should be returned or not.}
}
\value{
A tibble by default, otherwise a character vector.
}
\description{
Scrape a student's major from their transcript
}
\examples{
\dontrun{
scrape_major(dat)
}
}
