% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_sat.R
\name{scrape_sat}
\alias{scrape_sat}
\title{Scrape SAT scores from a student's transcript}
\usage{
scrape_sat(.data)
}
\arguments{
\item{.data}{A character vector returned from the read_transcript function.}
}
\value{
A tibble of SAT test scores
}
\description{
Scrape SAT scores from a student's transcript
}
\examples{
scrape_sat(dat)
}
