% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_satr.R
\name{scrape_satr}
\alias{scrape_satr}
\title{Scrape SATR scores from a student's transcript}
\usage{
scrape_satr(.data)
}
\arguments{
\item{.data}{A character vector returned from the read_transcript function.}
}
\value{
A tibble of SATR test scores
}
\description{
Scrape SATR scores from a student's transcript
}
\examples{
scrape_satr(dat)
}
