% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytranscript.R
\name{tidytranscript}
\alias{tidytranscript}
\title{Scrape all desired data from a student's transcript}
\usage{
tidytranscript(file)
}
\arguments{
\item{file}{A path to a file (either a single string or a raw vector)}
}
\value{
A tibble with all of the data that we care about
}
\description{
Scrape all desired data from a student's transcript
}
\examples{
\dontrun{
tidytranscript('data_raw/my_transcript.pdf')
}
}
